First part of major refactoring and Swift porting
- Refactored away almost all table cell subclasses
- Created SUSURLConnection for doing asynchronous NSURLConnections to the Subsonic server using blocks for callbacks
- Fixed player button not showing in CustomUITableViewControllers
- Ported CustomUITableView(Cell), other cells, and CellOverlay to Swift
- Ported CustomUIAlertView to Swift
- Ported FoldersViewController to Swift
- Ported ModalAlbumArtViewController to Swift
- Ported StoreViewController to Swift

+++ Classes/FolderDropdownDelegate.h
@@ -10,7 +10,7 @@
 @protocol FolderDropdownDelegate <NSObject>
 
 @required
-- (void)folderDropdownMoveViewsY:(float)y;
+- (void)folderDropdownMoveViewsY:(CGFloat)y;
 - (void)folderDropdownViewsFinishedMoving;
 - (void)folderDropdownSelectFolder:(NSNumber *)folderId;
